HWMonitor - back to fixed temp values
git-svn-id: http://subversion.assembla.com/svn/fakesmc/HWSensors4@683 958d8502-8caa-4642-8a4d-47b05a98fa7a

+++ HWMonitor/HWMonitorSensor.m
@@ -102,8 +102,8 @@
                 encoded = [HWMonitorSensor swapBytes:encoded];
                 
                 if ([type isEqualToString:@TYPE_SP78]){
-                    float v = ((float) encoded )/ 256.0f; //2^12
-                    return [[NSString alloc] initWithFormat:@"%2.1f°",v];
+                    int v = ((float) encoded )/ 256.0f; //2^12
+                    return [[NSString alloc] initWithFormat:@"%2d°",v];
                 }
